Stop using DOCTYPE to parse as XML.
Turns out that in the real world, only the Content Type header is used to determine the parsing engine used by the browser. Hoxy must mimic this behavior because that is what the author tested against. Not doing so means we cause undesirable changes to documents which have an XML doctype but are not syntactically valid as XML. In these cases, more information is preserved if we parse and render as HTML.